Testing out changing how stats work.
Before they checked if you passed the requirement, if you did you succeed.  Then rolled a random difficulty, and checked that against your stat.

I've changed it to more resemble d&d.  There is no way to be guaranteed a success now.  The difficulty is set, and you roll a random value and add floor((STAT-10)/2).  I also want to look at making skills a simple toggle, that gives advantage, also similar to d&d

The goal of this is to make stats less of a pass or maybe fail to always a maybe fail that someone with high stats fails less.

Conflicts:
	code/game/machinery/doors/airlock.dm
	code/modules/multiz/movement.dm